requirements.txt
# Core AWS SDK
boto3==1.34.0
botocore==1.34.0

# ETL and Data Processing
pandas==2.1.4
numpy==1.24.3
pyarrow==14.0.1
pyspark==3.5.0

# Database Connectors
cx_Oracle==8.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# API Integration
requests==2.31.0
tableau-api-lib==0.0.1

# AWS Glue
aws-glue-libs==4.0.0

# Monitoring and Logging
cloudwatch-logs==0.1.0
watchtower==3.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
moto==4.2.14

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Search and ML
scikit-learn==1.3.2
elasticsearch==8.11.0
